feat: Add Glossary for OBv3
The examples, in https://github.com/educredentials/obv3-examples
use IDs in contexts that are links. They should resolve to
a place that has some documentation. And that in future could be made
machine readable (e.g. rdf).

For now, we serve a simple HTML file. This file can be edited by hand.
We can link to it with #anchors.