chore: Ensure we don't wipe the database between docker-runs.
Wiping the database is:
- slow. It adds over a minute to the boot process of a docker compose up
- cumbersome. We start with a blank slate after every boot, which means
  issued badges, accounts, logins, etc all are gone.

Ideally the blank state is useful in itself, but it's lacking data and
setup currently. So after booting we need quite a few manual tasks each
time, to get relevant badges into users' backpack.

This feature was commented out in the commit:

  commit 2916276c9770cf18a4ea72373474ea6ce03e8b80
  Author: Okke Harsta <oharsta@zilverline.com>
  Date:   Wed Aug 18 14:24:39 2021 +0200

      Temporary comment to fix deployment

  Δ apps/mainsite/management/commands/seed.py

But, as "temporary fixes" tend to do, it remained in, and now causes
work and delays when developing.

We also add a persistent volume to docker-compose so that the data for
the mysql database is stored outside of the container and can live
across container lifecycles.