WIP: let seed command fail with errors when it fails
Currenly, failing seeds would be "eaten" by the try/except clauses, and
a calling script, e.g. the entrypoint.sh would think a seed was
performed successfully (no errors, exit code 0) and continue.
But that would start a server that does not have required data.

This change ensures errors are propagated to the caller of the script
and exit code is nonzero if a seed fails.

WIP, because we need to consider potential issues for production and
deployments that this might cause. If any.