feat: Change "sphereon" issuer to issuer with authorization flow
For this, we need to change some ENV vars.

We also need to differentiate between the request payload quite a lot,
so we introduced two distinct serializes each with their own model for
each type of request.

We then need to provide user-data with the authorization request, so
that we provide the issuer with the nessecary details to determine if we
are dealing with the expected user on issuance (the actual
authorization).