fix: Let local dev in docker-compose run the devserver
It tried to run "entrypoint.sh", whith the first argument
"entrypoint.sh" (itself). And the entrypoint executed the first
argument. Ergo: it caused an immediate stack-depth exception.

Running gunicorn in dev is impractical as it cannot easily reload etc.
So we use the django devserver as before.

To clarify what files are used, moved them into docker dir and named
them "dev" or "k8s".