ci: Don't build on PRs.
Building on PRs would release a new :latest image that would then be
deployed to k8s dev. We only want to build merged and accepted PRs.

If we were to build images on PRs, e.g. for a preview environment, we
should tag them differently, but that's for later