feat: Introduce tests for OidcClient
For now we only test EduIdUserInfo
This sets us up for the more complex business logic that requires
picking preferred names in future commits

Existing test in this module failed on attempting to create a new
database but had no permissions to do so. These tests were empty
scaffolding, so I simply switched to a superclass that does not try to
create databases for now.