fix: Remove proxy endpoints for .well known and other public oid4vci paths
These were broken and we don't need them.

The ProxyView class is expected to have a "queryset", the authentication
class dictates this. But it doesnt have one:

```
Exception Type: AssertionError at /ob3/.well-known/openid-credential-issuer
Exception Value: Cannot apply DjangoModelPermissionsOrAnonReadOnly on a view that does not set `.queryset` or have a `.get_queryset()` method.
```

Fixing this is not that hard, but unneeded, since we don't use nor need
these endpoints anyway. Clients, and wallets, use the issuer's
endpoints, not our backends'.