fix: Remove referer header requirement from login view
As a user on playground or a user recieving a mail, when I follow
a link in that mail, then I expect to be dropped into the login flow
immediately. So that I can skip the step where I have to choose between
"staff", "collection" and "backpack".

By removing the referer, we can link to the login link that will start
the authentication flow and will redirect the user into the
openidconnect flow immediately.
Currently this referer was required and caused an exception if the
header did not exist.