fix: Set userids to predictable eppn
This allows us to use a persistent id for users so that when they
log in via surfconext, we can find the appropriate socialaccount and
user.

Previously, EPPNs and uids differd between the ones in the seed and the
ones that surfconext uses.
When a user then logs in via surfconext, they don't see any badges in
their backpack