feat: Add filter for badgeclass entity_id on badge_instances
So that we can filter students' badge instances on "badgeclass".

This follows the existing pattern of server side filtering. But we don't
employ that consistently yet. Most filtering happens on the client-side.
Where the graphql simply returns the entire set and the client filters
on attributes.
But we do have existing server-side filtering. Mostly on booleans, a
date and quite a lot on ids.

By moving this to the server side we make the UI filtering a lot
simpler.

The UI part that uses this query will be implemented in a separate PR on
the UI repo.